requirements.txt
trendspy
praw
python-dotenv
pyyaml
loguru
anthropic
# urllib3
# ==1.25.11
